feat: show the diagram drag and drop on the new route
Refs: oph-1056
