+++ app/components/process/wizard.js
@@ -36,7 +36,6 @@ export default class ProcessWizard extends Component {
     UPLOAD_FILES: 'upload_files',
     SELECT_MAIN_PROCESS: 'select_main_process',
     CHANGE_MAIN_PROCESS: 'change_main_process',
-    STRUCTURE_DIAGRAMS: 'structure_diagrams',
     CREATE_PROCESS: 'create_process',
     UPDATE_PROCESS: 'update_process',
     CREATE_DIAGRAM_VERSION: 'create_diagram_version',
@@ -98,16 +97,6 @@ export default class ProcessWizard extends Component {
         canGoToNextStep: this.currentAction,
         nextStepButtonLabel: null,
         action: async () => await this.prepareWizard(),
-      },
-      {
-        step: this.wizardStep.STRUCTURE_DIAGRAMS,
-        title: 'Diagrammen structuur',
-        isStepShown: this.currentAction === WizardAction.STRUCTURE_DIAGRAMS,
-        canGoToNextStep: true,
-        nextStepButtonLabel: 'Aanpassen',
-        canCancelStep: true,
-        actionAtEndOfStep: async () =>
-          await this.saveDiagramStructure(this.diagramList),
       },
       {
         step: this.wizardStep.UPLOAD_FILES,
@@ -235,21 +224,6 @@ export default class ProcessWizard extends Component {
         this.disabledActions = [WizardAction.CHANGE_MAIN_PROCESS];
       }
     }
-  }
-
-  async saveDiagramStructure(diagramList) {
-    for (const main of diagramList.diagrams) {
-      await main.save();
-    }
-
-    const subItems = diagramList.diagrams.flatMap(
-      (main) => main.subItems ?? [],
-    );
-    for (const sub of subItems) {
-      await sub.save();
-    }
-
-    await diagramList.save();
   }
 
   async saveFileInDatabase(uploadedFile) {
